remove llvm toolchain
Since the beginning, we have added llvm 8.0.0 targeting ubuntu 16.04
release as an alternative to gcc.
However, this toolchain has not been popular among our users and
therefore, has not receive a lot of love. The current version is fairly
out of date vs the latest release.

Let's remove this toolchain entirely and guide folks who are interested
in LLVM toolchain toward a better alternative:
https://github.com/bazel-contrib/toolchains_llvm