Prepare for 0.0.1 release
In BzlMod, we have a top level module named `toolchains_buildbuddy`,
which contains the `buildbuddy` extension.
The extension will then be used to create the `buildbuddy_toolchain`
repository, which contains our toolchain and platform defintions.

Add automatic toolchains and exec platforms registration in the
MODULE.bazel of the `buildbuddy_toolchain` module. This means that any
repository which use our module would immediately get the relevant
toolchains and exec platforms available to them.

Fix documentations and comments in README.md and bzlmod examples.