cc_toolchain_config: coverage_feature stop providing profile
We have 3 features in this toolchain which are currently providing
"profile":
- coverage_feature
- gcc_coverage_map_format_feature
- llvm_coverage_map_format_feature

which we append all of them to the final feature list.

It's possible for multiple feature to be active at the same time when
user run `bazel coverage`. When this happens, the user would get this
error (reformatted by me for easier reading):

  Error in configure_features:
    Symbol "profile" is provided by all of the following features:
      coverage
      gcc_coverage_map_format

Since we mostly expect Ubuntu GCC toolchain while using BuildBuddy
toolchain, we can assume that gcc_coverage_map_format would always be
dormant.

Remove the "provides" from "coverage" feature.